+++ src/inputs/autocomplete/autocomplete.component.ts
@@ -9,7 +9,9 @@ import {BaseRESTService, SelectListInterface} from 'ramster-ui-core'
 @Component({
 	selector: 'rui-autocomplete',
 	templateUrl: './autocomplete.template.html',
-	styleUrls: ['../base/baseInput.styles.css']
+	styleUrls: [
+		'../base/baseInput.styles.css'
+	]
 })
 export class AutocompleteComponent extends BaseInputComponent {
 	@Input()