+++ src/validators/arrayNotEmpty.validator.ts
@@ -1,0 +1,5 @@
+import {FormControl} from '@angular/forms'
+
+export function arrayNotEmpty(control: FormControl): {[x: string]: any} | null {
+	return (control.value instanceof Array) && control.value.length ? null : {invalid: true}
+}