+++ src/validators/index.ts
@@ -1,5 +1,7 @@
+import {arrayNotEmpty} from './arrayNotEmpty.validator'
 import {checkEmailInUse} from './checkEmailInUse.validator'
 import {matchSibling} from './matchSibling.validator'
+import {objectNotEmpty} from './objectNotEmpty.validator'
 import {validateEmail} from './validateEmail.validator'
 
-export default {checkEmailInUse, matchSibling, validateEmail}
+export default {arrayNotEmpty, checkEmailInUse, matchSibling, objectNotEmpty, validateEmail}