+++ test/app/pages/home/home.component.ts
@@ -2,7 +2,7 @@
 
 import {ActivatedRoute} from '@angular/router'
 import {Component} from '@angular/core'
-import {FormControl} from '@angular/forms'
+import {FormControl, Validators} from '@angular/forms'
 
 import {
 	AutocompleteFieldDataInterface,
@@ -12,16 +12,19 @@ import {
 	InputFieldDataInterface,
 	SelectFieldDataInterface,
 	SlideToggleFieldDataInterface,
-	TextareaFieldDataInterface
+	TextareaFieldDataInterface,
+	validators
 } from '../../../../src'
 import {BasePageComponent, GlobalEventsService} from 'ramster-ui-core'
+
+import {TestModelRESTService} from '../../models/test/test.restService'
 
 @Component({
 	selector: 'app-page',
 	templateUrl: './home.template.pug',
 	styleUrls: [
 		'./home.styles.scss'
-	],
+	]
 })
 export class HomePageComponent extends BasePageComponent {
 	testAutocompleteFieldData: AutocompleteFieldDataInterface
@@ -36,7 +39,8 @@ export class HomePageComponent extends BasePageComponent {
 
 	constructor(
 		activatedRoute: ActivatedRoute,
-		globalEventsService: GlobalEventsService
+		globalEventsService: GlobalEventsService,
+		public testModelRESTService: TestModelRESTService
 	) {
 		super(activatedRoute, globalEventsService, ['reset'], ['onInitialDataLoaded'])
 	}
@@ -45,44 +49,47 @@ export class HomePageComponent extends BasePageComponent {
 		super.reset()
 
 		this.testAutocompleteFieldData = {
-			inputFormControl: new FormControl(''),
+			inputFormControl: new FormControl('', [Validators.required]),
+			loadSelectListOnInit: true,
 			placeholder: 'Autocomplete Input',
-			selectList: [{text: 'Value 1', value: 1}, {text: 'Value 2', value: 2}]
+			searchBoxValidators: [Validators.required],
+			selectList: [],
+			selectListRESTService: this.testModelRESTService
 		}
 		this.testAutocompleteWithChipsFieldData = {
 			hasChips: true,
-			inputFormControl: new FormControl([]),
+			inputFormControl: new FormControl([], [validators.arrayNotEmpty]),
 			placeholder: 'Autocomplete Input With Chips',
 			selectList: [{text: 'Value 1', value: 1}, {text: 'Value 2', value: 2}]
 		}
 		this.testCheckboxFieldData = {
-			inputFormControl: new FormControl(''),
+			inputFormControl: new FormControl('', [Validators.required]),
 			placeholder: 'Checkbox'
 		}
 		this.testDatepickerFieldData = {
-			inputFormControl: new FormControl(''),
+			inputFormControl: new FormControl('', [Validators.required]),
 			placeholder: 'Datepicker'
 		}
 		this.testFileInputFieldData = {
-			inputFormControl: new FormControl(''),
+			inputFormControl: new FormControl('', [Validators.required]),
 			placeholder: 'File Input'
 		}
 		this.testInputFieldData = {
-			inputFormControl: new FormControl(''),
+			inputFormControl: new FormControl('', [Validators.required]),
 			placeholder: 'Regular Input',
 			type: 'text'
 		}
 		this.testSelectFieldData = {
-			inputFormControl: new FormControl(''),
+			inputFormControl: new FormControl('', [Validators.required]),
 			placeholder: 'Select',
 			selectList: [{text: 'Option 1', value: 1}, {text: 'Option 2', value: 2}, {text: 'Option 3', value: 3}]
 		}
 		this.testSlideToggleFieldData = {
-			inputFormControl: new FormControl(''),
+			inputFormControl: new FormControl('', [Validators.required]),
 			placeholder: 'Slide Toggle'
 		}
 		this.testTextAreaFieldData = {
-			inputFormControl: new FormControl(''),
+			inputFormControl: new FormControl('', [Validators.required]),
 			placeholder: 'Text Area'
 		}
 		this.globalEventsService.setLayoutData({hasHeader: true})